+++ changelog.md
@@ -1,1 +1,3 @@
-* Añadido idioma ucraniano
+* Corregidos errores.
+
+* Añadida la posibilidad de hacer y restaurar copias de seguridad.